Organize test classes for qti tests
